+++ gimmemotifs/maelstrom/moap.py
@@ -1,16 +1,13 @@
-# Copyright (c) 2016 Simon van Heeringen <simon.vanheeringen@gmail.com>
-#
-# This module is free software. You can redistribute it and/or modify it under
-# the terms of the MIT License, see the file COPYING included with this
-# distribution.
 """ Module for motif activity prediction """
 import logging
 import os
+import warnings
 
 import numpy as np
 import pandas as pd
 from scipy.stats import hypergeom, mannwhitneyu
 from sklearn.ensemble import RandomForestClassifier
+from sklearn.exceptions import ConvergenceWarning
 from sklearn.linear_model import BayesianRidge, MultiTaskLassoCV
 from sklearn.multiclass import OneVsRestClassifier
 from sklearn.multioutput import MultiOutputRegressor
@@ -22,9 +19,7 @@ from tqdm.auto import tqdm
 
 from gimmemotifs import __version__
 from gimmemotifs.config import MotifConfig
-from gimmemotifs.motif import read_motifs
 from gimmemotifs.scanner import scan_regionfile_to_table
-from gimmemotifs.utils import pfmfile_location
 
 try:
     import xgboost  # noqa: optional
@@ -120,7 +115,7 @@ class BayesianRidgeMoap(Moap):
     supported_tables = ["score", "count"]
     ptype = "regression"
 
-    def __init__(self, scale=True, ncpus=None):
+    def __init__(self, scale=True, ncpus=None, disable_tqdm=False, *args, **kwargs):
         """Predict motif activities using Bayesian Ridge Regression.
 
         Parameters
@@ -141,6 +136,7 @@ class BayesianRidgeMoap(Moap):
             ncpus = int(MotifConfig().get_default_params().get("ncpus", 2))
         self.ncpus = ncpus
         self.scale = scale
+        self.disable_tqdm = disable_tqdm
 
     def fit(self, df_X, df_y):
         logger.info("Fitting BayesianRidge")
@@ -162,7 +158,7 @@ class BayesianRidgeMoap(Moap):
         model = BayesianRidge()
         logger.debug("Fitting model")
         coefs = []
-        for col in tqdm(y.columns, total=len(y.columns)):
+        for col in tqdm(y.columns, total=len(y.columns), disable=self.disable_tqdm):
             model.fit(X, y[col])
             coefs.append(model.coef_)
         logger.info("Done")
@@ -178,7 +174,15 @@ class XgboostRegressionMoap(Moap):
     supported_tables = ["score", "count"]
     ptype = "regression"
 
-    def __init__(self, scale=True, ncpus=None, random_state=None):
+    def __init__(
+        self,
+        scale=True,
+        ncpus=None,
+        disable_tqdm=False,
+        random_state=None,
+        *args,
+        **kwargs,
+    ):
         """Predict motif activities using XGBoost.
 
         Parameters
@@ -206,6 +210,7 @@ class XgboostRegressionMoap(Moap):
         self.ncpus = ncpus
         self.scale = scale
         self.random_state = random_state
+        self.disable_tqdm = disable_tqdm
 
     def fit(self, df_X, df_y):
         logger.info("Fitting XGBoostRegression")
@@ -239,7 +244,7 @@ class XgboostRegressionMoap(Moap):
         self.act_ = pd.DataFrame(index=X.columns)
 
         # Fit model
-        for col in tqdm(y.columns):
+        for col in tqdm(y.columns, disable=self.disable_tqdm):
             xgb.fit(X, y[col].values)
             d = xgb.get_booster().get_fscore()
             self.act_[col] = [d.get(m, 0) for m in X.columns]
@@ -397,7 +402,7 @@ class RFMoap(Moap):
     supported_tables = ["score", "count"]
     ptype = "classification"
 
-    def __init__(self, ncpus=None, random_state=None):
+    def __init__(self, ncpus=None, random_state=None, *args, **kwargs):
         """Predict motif activities using a random forest classifier
 
         Parameters
@@ -466,7 +471,7 @@ class MultiTaskLassoMoap(Moap):
     supported_tables = ["score", "count"]
     ptype = "regression"
 
-    def __init__(self, scale=True, ncpus=None, random_state=None):
+    def __init__(self, scale=True, ncpus=None, random_state=None, *args, **kwargs):
         """Predict motif activities using MultiTaskLasso.
 
         Parameters
@@ -543,7 +548,7 @@ class SVRMoap(Moap):
     supported_tables = ["score", "count"]
     ptype = "regression"
 
-    def __init__(self, scale=True, ncpus=None, random_state=None):
+    def __init__(self, scale=True, ncpus=None, random_state=None, *args, **kwargs):
         """Predict motif activities using Support Vector Regression.
 
         Parameters
@@ -591,7 +596,9 @@ class SVRMoap(Moap):
         clf = LinearSVR(random_state=self.random_state, dual="auto")
         self.model = MultiOutputRegressor(clf, n_jobs=1)
         logger.debug("Fitting model")
-        self.model.fit(df_X, df_y)
+        with warnings.catch_warnings():
+            warnings.simplefilter("ignore", category=ConvergenceWarning)
+            self.model.fit(df_X, df_y)
         logger.info("Done")
 
         self.act_ = pd.DataFrame(
@@ -622,57 +629,57 @@ def moap(
 ):
     """Run a single motif activity prediction algorithm.
 
-     Parameters
-     ----------
-     inputfile : str
-         :1File with regions (chr:start-end) in first column and either cluster
-         name in second column or a table with values.
+    Parameters
+    ----------
+    inputfile : str
+        File with regions (chr:start-end) in first column and either cluster
+        name in second column or a table with values.
 
-     method : str, optional
-         Motif activity method to use. Any of
-         'bayesianridge', 'xgboost', 'mwu', 'hypergeom',
-         'rf', 'multitasklasso', 'svr'. Default is 'hypergeom'.
+    method : str, optional
+        Motif activity method to use. Any of
+        'bayesianridge', 'xgboost', 'mwu', 'hypergeom',
+        'rf', 'multitasklasso', 'svr'. Default is 'hypergeom'.
 
-     scoring:  str, optional
-         Either 'score' or 'count'
+    scoring:  str, optional
+        Either 'score' or 'count'
 
-     outfile : str, optional
-         Name of outputfile to save the fitted activity values.
+    outfile : str, optional
+        Name of outputfile to save the fitted activity values.
 
-     motiffile : str, optional
-         Table with motif scan results. First column should be exactly the same
-         regions as in the inputfile.
+    motiffile : str, optional
+        Table with motif scan results. First column should be exactly the same
+        regions as in the inputfile.
 
-     pfmfile : str, optional
-         File with motifs in pfm format. Required when motiffile is not
-         supplied.
+    pfmfile : str, optional
+        File with motifs in pfm format. Required when motiffile is not
+        supplied.
 
-     genome : str, optional
-         Genome name, as indexed by gimme. Required when motiffile is not
-         supplied.
+    genome : str, optional
+        Genome name, as indexed by gimme. Required when motiffile is not
+        supplied.
 
-     zscore : bool, optional
-         Use z-score normalized motif scores.
+    zscore : bool, optional
+        Use z-score normalized motif scores.
 
-     gc : bool, optional
-         Equally distribute GC percentages in background sequences.
+    gc : bool, optional
+        Equally distribute GC percentages in background sequences.
 
-     subsample : float, optional
-         Fraction of regions to use.
+    subsample : float, optional
+        Fraction of regions to use.
 
-     random_state : numpy.random.RandomState object, optional
-         make predictions deterministic (where possible).
+    random_state : numpy.random.RandomState object, optional
+        make predictions deterministic (where possible).
 
-     ncpus : int, optional
-         Number of threads to use.
-         Default is the number specified in the config.
+    ncpus : int, optional
+        Number of threads to use.
+        Default is the number specified in the config.
 
     progress : bool or None, optional
-         provide progress bars for long computations.
+        provide progress bars for long computations.
 
-     Returns
-     -------
-     pandas DataFrame with motif activity
+    Returns
+    -------
+    pandas DataFrame with motif activity
     """
 
     if scoring and scoring not in ["score", "count"]:
@@ -685,7 +692,12 @@ def moap(
     else:
         df = pd.read_table(inputfile, index_col=0, comment="#")
 
-    clf = Moap.create(method, ncpus=ncpus, random_state=random_state)
+    disable_tqdm = not progress
+    if progress is None:
+        disable_tqdm = None
+    clf = Moap.create(
+        method, ncpus=ncpus, random_state=random_state, disable_tqdm=disable_tqdm
+    )
 
     if clf.ptype == "classification":
         if df.shape[1] != 1:
@@ -695,21 +707,10 @@ def moap(
             raise ValueError(f"columns should all be numeric for {method}")
 
     if motiffile is None:
-        if genome is None:
-            raise ValueError("need a genome")
-
-        pfmfile = pfmfile_location(pfmfile)
-        try:
-            _ = read_motifs(pfmfile)
-        except Exception:
-            logger.error(f"can't read motifs from {pfmfile}")
-            raise
-
         # scan for motifs
-        motif_names = [m.id for m in read_motifs(pfmfile)]
         if method == "classic" or scoring == "count":
             logger.info("motif scanning (counts)")
-            scores = scan_regionfile_to_table(
+            motifs = scan_regionfile_to_table(
                 inputfile,
                 genome,
                 "count",
@@ -722,7 +723,7 @@ def moap(
             )
         else:
             logger.info("motif scanning (scores)")
-            scores = scan_regionfile_to_table(
+            motifs = scan_regionfile_to_table(
                 inputfile,
                 genome,
                 "score",
@@ -733,8 +734,6 @@ def moap(
                 random_state=random_state,
                 progress=progress,
             )
-        motifs = pd.DataFrame(scores, index=df.index, columns=motif_names)
-
     elif isinstance(motiffile, pd.DataFrame):
         motifs = motiffile
     else:
@@ -751,11 +750,13 @@ def moap(
             return out
 
     if subsample is not None:
-        n = int(subsample * df.shape[0])
-        logger.debug(f"Subsampling {n} regions")
-        df = df.sample(n, random_state=random_state)
+        n = int(subsample * motifs.shape[0])
+        logger.debug(f"Subsampling to {n} regions")
+        motifs = motifs.sample(n, random_state=random_state)
 
-    motifs = motifs.loc[df.index]
+    # subset df if motifs was subsampled, or if regions from the inputfile
+    # could not be found in the genome (e.g. scaffolds, alt regions, etc.)
+    df = df.loc[motifs.index]
 
     clf.fit(motifs, df)